Working Qt interface (debuggers broken)
