+++ Qt/QtHost.cpp
@@ -2,6 +2,7 @@
 
 #include <QFileInfo>
 #include <QDebug>
+#include <QDir>
 
 #include "QtHost.h"
 #include "LogManager.h"
@@ -15,6 +16,9 @@
 #include "ui/ui_context.h"
 #include "GPU/ge_constants.h"
 #include "EmuThread.h"
+#include "UI/GameInfoCache.h"
+
+const char *stateToLoad = NULL;
 
 std::string boot_filename = "";
 Texture *uiTexture;
@@ -26,6 +30,11 @@ std::string game_title;
 
 event m_hGPUStepEvent;
 recursive_mutex m_hGPUStepMutex;
+
+recursive_mutex pendingMutex;
+static bool isMessagePending;
+static std::string pendingMessage;
+static std::string pendingValue;
 
 QtHost::QtHost(MainWindow *mainWindow_)
     : mainWindow(mainWindow_)
@@ -213,6 +222,106 @@ void QtHost::NextGPUStep()
 	m_hGPUStepEvent.notify_one();
 }
 
+void NativeInit(int argc, const char *argv[], const char *savegame_directory, const char *external_directory, const char *installID)
+{
+	std::string config_filename;
+	Common::EnableCrashingOnCrashes();
+	isMessagePending = false;
+
+	std::string user_data_path = savegame_directory;
+
+	VFSRegister("", new DirectoryAssetReader("assets/"));
+	VFSRegister("", new DirectoryAssetReader(user_data_path.c_str()));
+
+	config_filename = user_data_path + "ppsspp.ini";
+
+	g_Config.Load(config_filename.c_str());
+
+	const char *fileToLog = 0;
+
+	bool hideLog = true;
+#ifdef _DEBUG
+	hideLog = false;
+#endif
+
+	bool gfxLog = false;
+	// Parse command line
+	LogTypes::LOG_LEVELS logLevel = LogTypes::LINFO;
+	for (int i = 1; i < argc; i++) {
+		if (argv[i][0] == '-') {
+			switch (argv[i][1]) {
+			case 'd':
+				// Enable debug logging
+				logLevel = LogTypes::LDEBUG;
+				break;
+			case 'g':
+				gfxLog = true;
+				break;
+			case 'j':
+				g_Config.bJit = true;
+				g_Config.bSaveSettings = false;
+				break;
+			case 'i':
+				g_Config.bJit = false;
+				g_Config.bSaveSettings = false;
+				break;
+			case 'l':
+				hideLog = false;
+				break;
+			case 's':
+				g_Config.bAutoRun = false;
+				g_Config.bSaveSettings = false;
+				break;
+			case '-':
+				if (!strcmp(argv[i], "--log") && i < argc - 1)
+					fileToLog = argv[++i];
+				if (!strncmp(argv[i], "--log=", strlen("--log=")) && strlen(argv[i]) > strlen("--log="))
+					fileToLog = argv[i] + strlen("--log=");
+				if (!strcmp(argv[i], "--state") && i < argc - 1)
+					stateToLoad = argv[++i];
+				if (!strncmp(argv[i], "--state=", strlen("--state=")) && strlen(argv[i]) > strlen("--state="))
+					stateToLoad = argv[i] + strlen("--state=");
+				break;
+			}
+		}
+		else if (fileToStart.isNull())
+		{
+			fileToStart = QString(argv[i]);
+			if (!QFile::exists(fileToStart))
+			{
+				qCritical("File '%s' does not exist!", qPrintable(fileToStart));
+				exit(1);
+			}
+		}
+		else
+		{
+			qCritical("Can only boot one file. Ignoring file '%s'", qPrintable(fileToStart));
+		}
+	}
+
+	if (g_Config.currentDirectory == "")
+	{
+		g_Config.currentDirectory = QDir::homePath().toStdString();
+	}
+
+	g_Config.memCardDirectory = QDir::homePath().toStdString()+"/.ppsspp/";
+	g_Config.flashDirectory = g_Config.memCardDirectory+"/flash/";
+
+	LogManager::Init();
+	if (fileToLog != NULL)
+		LogManager::GetInstance()->ChangeFileLog(fileToLog);
+
+	LogManager::GetInstance()->SetLogLevel(LogTypes::G3D, LogTypes::LERROR);
+
+	g_gameInfoCache.Init();
+
+#if !defined(USING_GLES2)
+	// Start Desktop UI
+	MainWindow* mainWindow = new MainWindow();
+	mainWindow->show();
+#endif
+}
+
 int NativeMix(short *audio, int num_samples)
 {
 	if (g_mixer)
@@ -288,8 +397,26 @@ void NativeRender()
 	screenManager->render();
 }
 
+void NativeMessageReceived(const char *message, const char *value)
+{
+	lock_guard lock(pendingMutex);
+	if (!isMessagePending) {
+		pendingMessage = message;
+		pendingValue = value;
+		isMessagePending = true;
+	}
+}
+
 void NativeUpdate(InputState &input)
 {
+	{
+		lock_guard lock(pendingMutex);
+		if (isMessagePending) {
+			screenManager->sendMessage(pendingMessage.c_str(), pendingValue.c_str());
+			isMessagePending = false;
+		}
+	}
+
 	UIUpdateMouse(0, input.pointer_x[0], input.pointer_y[0], input.pointer_down[0]);
 	screenManager->update(input);
 }
